rename stock order component
