+++ src/app/modules/profile/components/profile/profile-wallets/profile-wallet-orders/profile-wallet-orders.config.ts
@@ -1,4 +1,4 @@
-import { GetOrdersRequest } from '../../../../../stock/components/stock/stock-order/stock-order.model';
+import { GetOrdersRequest } from '../../../../../stock/components/stock/stock-order-buy/stock-order-buy.model';
 
 export const defaultGetOrdersRequest: GetOrdersRequest = {
   wallet_id: '',
