+++ src/app/modules/stock/components/stock-analyse/stock-order-buy/stock-order-buy.component.html
@@ -8,8 +8,8 @@
   contentTemplate="contentTemplatePopup">
   <div *dxTemplate="let d of 'contentTemplatePopup'" class="w-100 h-100">
     <div class="row g-4 gx-0">
-      <h5 class="col-12 text-center" i18n="@@stock-order.Current-price">Current price: {{ cryptoDetails.market_data.current_price | number:'1.2-2' }}&nbsp;{{ selectedWallet.currency }}</h5>
-      <h6 class="col-12 text-center mt-0" i18n="@@stock-order.Free-wallet-funds">Free wallet funds: {{ selectedWallet.value | number:'1.2-2' }} {{ selectedWallet.currency }}</h6>
+      <h5 class="col-12 text-center" i18n="@@stock-order-buy.Current-price">Current price: {{ cryptoDetails.market_data.current_price | number:'1.2-2' }}&nbsp;{{ selectedWallet.currency }}</h5>
+      <h6 class="col-12 text-center mt-0" i18n="@@stock-order-buy.Free-wallet-funds">Free wallet funds: {{ selectedWallet.value | number:'1.2-2' }} {{ selectedWallet.currency }}</h6>
 
       @if (orderSide === OrderSide.Sell) {
         <div class="col-6">
@@ -21,7 +21,7 @@
             [(value)]="nominal"
             (onValueChanged)="onNominalChanged($event)"
             validationMessageMode="always"
-            i18n-label="@@stock-order.Nominal:Nominal">
+            i18n-label="@@stock-order-buy.Nominal:Nominal">
             <dx-validator>
               <dxi-validation-rule type="required"></dxi-validation-rule>
             </dx-validator>
@@ -36,7 +36,7 @@
             (onSelectionChanged)="onCryptoSelectionChanged()"
             label="Crypto"
             labelMode="floating"
-            i18n-label="@@stock-order.Crypto:Crypto"
+            i18n-label="@@stock-order-buy.Crypto:Crypto"
             valueExpr="currency_name"
             validationMessageMode="always"
             displayExpr="currency_name">
@@ -57,9 +57,9 @@
           format="#0.##"
           (onValueChanged)="onAmountChanged($event)"
           validationMessageMode="always"
-          i18n-label="@@stock-order.Amount:Amount">
+          i18n-label="@@stock-order-buy.Amount:Amount">
           <dx-validator>
-            <dxi-validation-rule type="custom" [validationCallback]="amountValidationCallback.bind(this)" message="Exceeded free wallet funds" i18n-message="@@stock-order.Exceeded-free-wallet-funds"></dxi-validation-rule>
+            <dxi-validation-rule type="custom" [validationCallback]="amountValidationCallback.bind(this)" message="Exceeded free wallet funds" i18n-message="@@stock-order-buy.Exceeded-free-wallet-funds"></dxi-validation-rule>
           </dx-validator>
         </dx-number-box>
       </div>
@@ -72,7 +72,7 @@
           (onSelectionChanged)="onWalletSelectionChanged()"
           label="Wallet"
           labelMode="floating"
-          i18n-label="@@stock-order.Wallet:Wallet"
+          i18n-label="@@stock-order-buy.Wallet:Wallet"
           valueExpr="id"
           validationMessageMode="always"
           displayExpr="currency"
@@ -95,7 +95,7 @@
             [(value)]="nominal"
             (onValueChanged)="onNominalChanged($event)"
             validationMessageMode="always"
-            i18n-label="@@stock-order.Nominal:Nominal">
+            i18n-label="@@stock-order-buy.Nominal:Nominal">
             <dx-validator>
               <dxi-validation-rule type="required"></dxi-validation-rule>
             </dx-validator>
@@ -110,7 +110,7 @@
             (onSelectionChanged)="onCryptoSelectionChanged()"
             label="Crypto"
             labelMode="floating"
-            i18n-label="@@stock-order.Crypto:Crypto"
+            i18n-label="@@stock-order-buy.Crypto:Crypto"
             valueExpr="currency_name"
             validationMessageMode="always"
             displayExpr="currency_name">
@@ -137,10 +137,10 @@
       <div class="col-12">
         <div class="row g-4">
           <div class="col-lg-auto col-6 ms-lg-auto">
-            <dx-button class="w-100" type="normal" text="Cancel" i18n-text="@@stock-order.Cancel" (onClick)="close()"></dx-button>
+            <dx-button class="w-100" type="normal" text="Cancel" i18n-text="@@stock-order-buy.Cancel" (onClick)="close()"></dx-button>
           </div>
           <div class="col-lg-auto col-6">
-            <dx-button class="w-100" type="success" text="Confirm order" i18n-text="@@stock-order.Confirm-order" (onClick)="confirmOrder()"></dx-button>
+            <dx-button class="w-100" type="success" text="Confirm order" i18n-text="@@stock-order-buy.Confirm-order" (onClick)="confirmOrder()"></dx-button>
           </div>
         </div>
       </div>
