+++ src/app/modules/stock/components/stock-analyse/stock-order-buy/stock-order-buy.config.ts
@@ -1,5 +1,5 @@
 import { ButtonType } from 'devextreme/common';
-import { orderAvailableTypes, OrderSide, orderSidesMapReverse, orderSidesStringFiatWalletOperationPrefixMap, orderSideTitles, OrderStatus, orderStatusesMapReverse, OrderType, OrderTypeDetail } from './stock-order.model';
+import { orderAvailableTypes, OrderSide, orderSidesMapReverse, orderSidesStringFiatWalletOperationPrefixMap, orderSideTitles, OrderStatus, orderStatusesMapReverse, OrderType, OrderTypeDetail } from './stock-order-buy.model';
 
 export const getOrderButtonTypeType = (orderType: OrderType, selectedOrderType: OrderType): ButtonType => {
   return orderType === selectedOrderType ? 'default' : 'normal';
@@ -12,15 +12,15 @@ export const getOrderAvailableTypes = (orderSide: OrderSide): OrderTypeDetail[] 
 }
 
 export const getPopupTitle = (orderSide: OrderSide): string => {
-  return orderSideTitles.get(orderSide) ?? $localize`:@@stock-order-config.Execute-market-order:Execute market order`;
+  return orderSideTitles.get(orderSide) ?? $localize`:@@stock-order-buy-config.Execute-market-order:Execute market order`;
 }
 
 export const getOrderHistoryEntrySideLabel = (orderSide: string): string => {
-  return OrderSide[orderSidesMapReverse.get(orderSide) as OrderSide] ?? $localize`:@@stock-order-config.Unknown:Unknown`;
+  return OrderSide[orderSidesMapReverse.get(orderSide) as OrderSide] ?? $localize`:@@stock-order-buy-config.Unknown:Unknown`;
 }
 
 export const getOrderHistoryEntryStatusLabel = (orderStatus: string): string => {
-  return OrderStatus[orderStatusesMapReverse.get(orderStatus) as OrderStatus] ?? $localize`:@@stock-order-config.Unknown:Unknown`;
+  return OrderStatus[orderStatusesMapReverse.get(orderStatus) as OrderStatus] ?? $localize`:@@stock-order-buy-config.Unknown:Unknown`;
 }
 
 export const getOrderHistoryEntryCashQuantityPrefixLabel = (orderSide: string): string => {
